refactor: clean code
CU-3cy5cnn
