+++ src/components/pages/index.js
@@ -1,0 +1,3 @@
+import SuccessPage from "./SuccessPage";
+
+export default SuccessPage;
