+++ src/helper/index.js
@@ -1,0 +1,1 @@
+export { insuranceFormSchema, beneficiaryFormSchema } from "./yupSchema";
